% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{addTz}
\alias{addTz}
\title{Function to calculate Tz from wave parameter data}
\usage{
addTz(df1, gamma, tpindex)
}
\arguments{
\item{input_df}{}

\item{variables}{= atomic vector of input variable names for pivot table}

\item{bins}{= atomic vector of bins for each of the variables}
}
\description{
Function to calculate Tz from wave parameter data
}
\examples{
generatePivot(df,c("Hm0","Tp"),bins = c())
}

