% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/generatePivot.r
\name{calcWavePower}
\alias{calcWavePower}
\title{A Metocean function}
\usage{
calcWavePower(Hs, Tp)
}
\arguments{
\item{Hs}{- Significant wave height [m], Tp - Peak wave period}
}
\description{
This function calculates wave power based on Hs and Tp
}
\examples{
calcWavePower()
}
\keyword{Metocean}

