% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{combineData}
\alias{combineData}
\title{Function to join time series of files into a single file}
\usage{
combineData(rootDir, timeFiles)
}
\arguments{
\item{rootDir}{string of the root directory of all the files}

\item{timeFiles}{vector of strings giving the file names}
}
\value{
df dataframe with date and variables
}
\description{
Function to join time series of files into a single file
}
\examples{
combineData(rootDir,timeFiles)
}

