% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{convertDVRSAsciiToMikeAscii}
\alias{convertDVRSAsciiToMikeAscii}
\title{Function to convert Mike extract to Ascii DVRS file to a standard format}
\usage{
convertDVRSAsciiToMikeAscii(df1)
}
\arguments{
\item{time}{series file}
}
\description{
Function to convert Mike extract to Ascii DVRS file to a standard format
}
\examples{
convertDVRSAsciiToMikeAscii(df1)
}

