% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{generatePivot}
\alias{generatePivot}
\title{Function to generate a pivot table of percentage occurances}
\usage{
generatePivot(df1, variables, bins, ...)
}
\arguments{
\item{df1}{Combined data frame}

\item{variables}{vecotr of named variables within data frame}

\item{bins}{vector of bins for each of the named variables}
}
\value{
pivot_table
}
\description{
Function to generate a pivot table of percentage occurances
}
\examples{
generatePivot(df1,variables,bins)
}

