% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/processNCEP.r
\name{plotAnnualWaveRose}
\alias{plotAnnualWaveRose}
\title{A Metocean function}
\usage{
plotAnnualWaveRose(nww3Param, wave_params, breaks)
}
\arguments{
\item{File}{name text}
}
\description{
This function reads in a Mike by DHI time series file, renames headers and brings into a DF
}
\examples{
plotAnnualWaveRose()
}
\keyword{Metocean}

