% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{plotRose}
\alias{plotRose}
\title{Function to plot wind and wave rose plots}
\usage{
plotRose(df, hs, wd, bins)
}
\arguments{
\item{df}{time series of metocean data}

\item{hs}{magnitude variable}

\item{bins}{number of bins in the}

\item{dm}{direction variable}
}
\description{
Function to plot wind and wave rose plots
}
\examples{
plotRose()
}

