% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/metocean.R
\name{plotScatter}
\alias{plotScatter}
\title{Function to plot wave scatter plot}
\usage{
plotScatter(input_df, col_index = c(2, 3, 4), ...)
}
\arguments{
\item{input_df}{}

\item{col_index}{Index of required variables for scatter plot}
}
\value{
multiple scatter plot objects
}
\description{
Function to plot wave scatter plot
}
\examples{
plotScatter(input_df,col_index)
}

