% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{plotSeasonalRose}
\alias{plotSeasonalRose}
\title{Function from openair to plot a wave rose from a time series}
\usage{
plotSeasonalRose(nww3Param)
}
\arguments{
\item{input_df_time_series}{}
}
\description{
Function from openair to plot a wave rose from a time series
}
\examples{
plotSeasonalRose()
}

