% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/metocean.R
\name{plotWaveParams}
\alias{plotWaveParams}
\title{Function to plot wave time series}
\usage{
plotWaveParams(nww3Param)
}
\arguments{
\item{input_df}{}
}
\value{
Time series of Hmo, Tp, and Dp
}
\description{
Function to plot wave time series
}
\examples{
plotScatter(input_df,col_index)
}

