% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{readMikeAsciiDVRS}
\alias{readMikeAsciiDVRS}
\title{Function to read Mike by DHI automatically generated time series}
\usage{
readMikeAsciiDVRS(data_file)
}
\arguments{
\item{time}{series file}
}
\description{
Function to read Mike by DHI automatically generated time series
}
\examples{
readMikeAscii(data_file)
}

