% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{readMikeDFS0}
\alias{readMikeDFS0}
\title{Function to read Mike by DHI text file into a data frame}
\usage{
readMikeDFS0(data_file)
}
\arguments{
\item{data_file}{ascii file with time series of metocean data}
}
\value{
data_frame
}
\description{
Function to read Mike by DHI text file into a data frame
}
\examples{
readMikeDFS0()
}
\keyword{metocean}

