% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/processNCEP.r
\name{waveRose}
\alias{waveRose}
\title{A Metocean function}
\usage{
waveRose(mydata, ws = "ws", wd = "wd", ws2 = NA, wd2 = NA, ws.int = 2,
  angle = 30, type = "default", bias.corr = TRUE, cols = "default",
  grid.line = NULL, width = 1, seg = NULL, auto.text = TRUE,
  breaks = 4, offset = 10, max.freq = NULL, paddle = TRUE,
  key.header = NULL, key.footer = "(m/s)", key.position = "bottom",
  key = TRUE, dig.lab = 5, statistic = "prop.count", pollutant = NULL,
  annotate = TRUE, border = NA, ...)
}
\arguments{
\item{File}{name text}
}
\description{
This function reads in a Mike by DHI time series file, renames headers and brings into a DF
}
\examples{
waveRose()
}
\keyword{Metocean}

