% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{addTz}
\alias{addTz}
\title{Function to calculate Tz from wave parameter data}
\usage{
addTz(df1, hsindex, tpindex)
}
\arguments{
\item{hsindex}{- index of Hs}

\item{tpindex}{- index of Tp}

\item{input_df}{}
}
\description{
Function to calculate Tz from wave parameter data
}
\examples{
addTz(df1,hsindex,tpindex)
}

