% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/generatePivot.r
\name{calcWavePower}
\alias{calcWavePower}
\title{A Cat Function}
\usage{
calcWavePower(Hs, Tp)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

