% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{generatePivot}
\alias{generatePivot}
\title{Function to generate a pivot table of percentage occurances}
\usage{
generatePivot(df1, variables, bins, ...)

generatePivot(df1, variables, bins, ...)
}
\arguments{
\item{df1}{Combined data frame}

\item{variables}{vecotr of named variables within data frame}

\item{bins}{vector of bins for each of the named variables}

\item{target}{the value variable}

\item{input_df}{}

\item{variables}{= atomic vector of input variable names for pivot table}

\item{bins}{= atomic vector of bins for each of the variables}
}
\value{
pivot_table
}
\description{
Function to generate a pivot table of percentage occurances

Function to get a pivot table of Hm0 Tp occurances
}
\examples{
generatePivot(df1,variables,target,bins)
generatePivot(df,c("Hm0","Tp"),bins = c())
}

