% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/generatePivot.r
\name{plotScatter}
\alias{plotScatter}
\title{Function to generate scatter plots for wave data}
\usage{
plotScatter(input_df, my_col_brew, col_index = c(2, 3, 4), col_hmo, col_tp,
  col_dir, ...)
}
\arguments{
\item{Dataframe}{of time series of wave data}
}
\value{
Returns and plot object
}
\description{
Plots scatter diagrams for wave parameters of specified
}
\examples{
plotScatter(input_df)
}
\keyword{processing}
\keyword{scatter,}
\keyword{waves,}

