% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/polarPlot.r
\name{readMikeDFS0}
\alias{readMikeDFS0}
\title{A Metocean function}
\usage{
readMikeDFS0(data_file)
}
\arguments{
\item{File}{name text}
}
\description{
This function reads in a Mike by DHI time series file, renames headers and brings into a DF
}
\examples{
readMikeDFS0()
}
\keyword{Metocean}

