% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metocean.R
\name{waveRose}
\alias{waveRose}
\title{Function from openair to plot a wave rose from a time series}
\usage{
waveRose(mydata, ws = "ws", wd = "wd", ws2 = NA, wd2 = NA, ws.int = 2,
  angle = 30, type = "default", bias.corr = TRUE, cols = "default",
  grid.line = NULL, width = 1, seg = NULL, auto.text = TRUE,
  breaks = 4, offset = 10, max.freq = NULL, paddle = TRUE,
  key.header = NULL, key.footer = "(m/s)", key.position = "bottom",
  key = TRUE, dig.lab = 5, statistic = "prop.count", pollutant = NULL,
  annotate = TRUE, border = NA, ...)
}
\arguments{
\item{input_df_time_series}{}
}
\description{
Function from openair to plot a wave rose from a time series
}
\examples{
waveRose()
}

